backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# OpenAI
openai==1.12.0

# Google APIs
google-auth==2.28.0
google-api-python-client==2.118.0

# Request handling
python-multipart==0.0.9
python-dotenv==1.0.1

# Data validation
pydantic==2.6.1
pydantic-settings==2.2.0

# HTTP client
httpx==0.26.0
